separated add dialog to a different file
