+++ src/components/contacts/ContactList.js
@@ -17,18 +17,13 @@ import InputLabel from "@mui/material/InputLabel";
 import Select from "@mui/material/Select";
 import MenuItem from "@mui/material/MenuItem";
 import { useHistory } from "react-router-dom";
+import { ContactsAddDialog } from "./ContactsAddDialog";
 
 export const ContactList = ({ dashboardView }) => {
     const [myContacts, setMyContacts] = useState([]);
     const [contactsTypes, setContactsTypes] = useState([]);
     const { getCurrentUser } = useSimpleAuth();
-    const [newContact, setNewContact] = useState({
-        contactsTypeId: "",
-        userId: getCurrentUser().id,
-        starred: false,
-    });
     const [expanded, setExpanded] = useState(false);
-    const [open, setOpen] = useState(false);
     const history = useHistory();
 
     useEffect(() => {
@@ -46,26 +41,8 @@ export const ContactList = ({ dashboardView }) => {
         );
     }, []);
 
-    const handleClickOpen = () => {
-        setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
-    };
-
     const handleChange = (panel) => (event, isExpanded) => {
         setExpanded(isExpanded ? panel : false);
-    };
-
-    const addContact = (event) => {
-        event.preventDefault();
-        handleClose();
-        ContactsRepository.addContact(newContact).then(() =>
-            ContactsRepository.findContactsByUser(getCurrentUser().id).then(
-                (data) => setMyContacts(data)
-            )
-        );
     };
 
     return (
@@ -73,129 +50,11 @@ export const ContactList = ({ dashboardView }) => {
             <Typography variant="h1" gutterBottom align="center" fontSize="3em">
                 My Contacts
             </Typography>
-            {history.location.pathname === "/" ? (
-                ""
-            ) : (
-            <Typography align="center">
-                <Button variant="contained" onClick={handleClickOpen} sx={{marginBottom: "2em"}}>
-                    Add Contact
-                </Button>
-            </Typography>
-            )}
-            <Dialog open={open} onClose={handleClose}>
-                <form onSubmit={addContact}>
-                    <DialogTitle>Add Contact</DialogTitle>
-                    <DialogContent>
-                        <TextField
-                            autoFocus
-                            margin="dense"
-                            id="name"
-                            label="Name"
-                            type="text"
-                            required
-                            fullWidth
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.name = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="phone"
-                            label="Phone Number"
-                            type="tel"
-                            required
-                            fullWidth
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.phoneNumber = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="street"
-                            label="Street Address"
-                            type="text"
-                            required
-                            fullWidth
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.addressStreet = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="city"
-                            label="City"
-                            required
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.addressCity = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="state"
-                            label="State"
-                            required
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.addressState = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="zipCode"
-                            label="Zip Code"
-                            required
-                            type="number"
-                            onChange={(event) => {
-                                const copy = { ...newContact };
-                                copy.addressZipCode = event.target.value;
-                                setNewContact(copy);
-                            }}
-                        />
-                        <FormControl required sx={{ m: 1, minWidth: 225 }}>
-                            <InputLabel id="species-label">
-                                Contact type
-                            </InputLabel>
-                            <Select
-                                labelId="contact-type-label"
-                                id="contact-type"
-                                value={newContact.contactsTypeId}
-                                label="contact-type"
-                                onChange={(event) => {
-                                    const copy = { ...newContact };
-                                    copy.contactsTypeId = parseInt(
-                                        event.target.value
-                                    );
-                                    setNewContact(copy);
-                                }}
-                            >
-                                {contactsTypes.map((contactType) => (
-                                    <MenuItem
-                                        key={`contactType--${contactType.id}`}
-                                        value={contactType.id}
-                                    >
-                                        {contactType.type}
-                                    </MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-                    <DialogActions>
-                        <Button onClick={handleClose}>Cancel</Button>
-                        <Button type="submit">Add</Button>
-                    </DialogActions>
-                </form>
-            </Dialog>
+            <ContactsAddDialog
+                userId={getCurrentUser().id}
+                setMyContacts={setMyContacts}
+                contactsTypes={contactsTypes}
+            />
             <Box>
                 {myContacts.map((contact) => {
                     return (
